res fixed(mostly blindly) some 64bit issues reported by Mike Gist.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22123 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/win32/winsyspath.cpp
@@ -98,7 +98,7 @@ static DWORD STDAPICALLTYPE MyGetLPN (LPCSTR lpszShortPath, LPSTR lpszLongPath,
     }
     nextpos = bs ? bs + 1 : 0;
   }
-  return (cchBuffer - bufRemain);
+  return (cchBuffer - (DWORD)bufRemain);
 #undef BUFCAT
 }
